% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_exists}
\alias{cache_exists}
\title{Check if cache exists for specific data}
\usage{
cache_exists(name = NULL, year = NULL, cache_dir = NULL)
}
\arguments{
\item{name}{Department code (optional)}

\item{year}{Academic year (optional)}

\item{cache_dir}{Base cache directory (optional)}
}
\value{
Logical indicating if cache exists
}
\description{
Check if cache exists for specific data
}
