% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorecourses-package.R
\docType{package}
\name{explorecourses-package}
\alias{explorecourses}
\alias{explorecourses-package}
\title{Stanford}
\description{
This package contains R code specific to the Stanford University
web applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coatless-rpkg/explorecourses}
  \item Report bugs at \url{https://github.com/coatless-rpkg/explorecourses/issues}
}

}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{james.balamuta@gmail.com} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID}) [copyright holder]

}
\keyword{internal}
