% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{extract_basic_course_info}
\alias{extract_basic_course_info}
\title{Extract basic course information}
\usage{
extract_basic_course_info(course)
}
\arguments{
\item{course}{XML node for a course}
}
\value{
Tibble with basic course information
}
\description{
Extract basic course information
}
\keyword{internal}
