% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{extract_section_data}
\alias{extract_section_data}
\title{Extract section data including schedules}
\usage{
extract_section_data(course, course_id)
}
\arguments{
\item{course}{XML node for a course}

\item{course_id}{Course ID for joining}
}
\value{
Tibble with section and schedule data
}
\description{
Extract section data including schedules
}
\keyword{internal}
