% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{extract_section_info}
\alias{extract_section_info}
\title{Extract section information}
\usage{
extract_section_info(section, course_id)
}
\arguments{
\item{section}{XML node for a section}

\item{course_id}{Course ID for joining}
}
\value{
Tibble with section information
}
\description{
Extract section information
}
\keyword{internal}
