% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{fetch_departments_raw}
\alias{fetch_departments_raw}
\title{Fetch raw department data from API}
\usage{
fetch_departments_raw()
}
\value{
Raw XML content
}
\description{
Fetch raw department data from API
}
\keyword{internal}
