% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{init_cache_dirs}
\alias{init_cache_dirs}
\title{Initialize cache directories}
\usage{
init_cache_dirs(cache_dir)
}
\arguments{
\item{cache_dir}{Base cache directory}
}
\value{
List of cache paths
}
\description{
Initialize cache directories
}
\keyword{internal}
