% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{make_api_request}
\alias{make_api_request}
\title{Make API request to ExploreCourses endpoint}
\usage{
make_api_request(url)
}
\arguments{
\item{url}{API endpoint URL}
}
\value{
Raw response content or error
}
\description{
Make API request to ExploreCourses endpoint
}
\keyword{internal}
