% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{parse_courses}
\alias{parse_courses}
\title{Parse course XML into a data frame}
\usage{
parse_courses(xml_content)
}
\arguments{
\item{xml_content}{XML content from fetch_department_courses}
}
\value{
A list of data frames containing course information
}
\description{
Parse course XML into a data frame
}
