% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{process_department}
\alias{process_department}
\title{Process a single department}
\usage{
process_department(name, cache_dir = NULL, p = NULL)
}
\arguments{
\item{name}{Department code}

\item{cache_dir}{Base cache directory}

\item{p}{Progress handler}
}
\value{
Data frame of parsed course information
}
\description{
Process a single department
}
\keyword{internal}
