% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_departments_xml}
\alias{process_departments_xml}
\title{Process departments XML data into a data frame}
\usage{
process_departments_xml(xml_doc)
}
\arguments{
\item{xml_doc}{XML document (xml2 doc object)}
}
\value{
Data frame of departments
}
\description{
Process departments XML data into a data frame
}
\keyword{internal}
