% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{read_cache}
\alias{read_cache}
\title{Read course data from cache}
\usage{
read_cache(cache_dir, dept = NULL)
}
\arguments{
\item{cache_dir}{Directory containing cached files}

\item{dept}{Department code (optional)}
}
\value{
Data frame of course data
}
\description{
Read course data from cache
}
