% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_academic_year}
\alias{validate_academic_year}
\title{Validate academic year format}
\usage{
validate_academic_year(year)
}
\arguments{
\item{year}{Academic year in format YYYYYYYY or NULL}
}
\value{
Validated year string or error
}
\description{
Validate academic year format
}
\keyword{internal}
