% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{write_cache}
\alias{write_cache}
\title{Write course data to cache}
\usage{
write_cache(data, cache_dir, dept)
}
\arguments{
\item{data}{Course data frame}

\item{cache_dir}{Directory to cache results}

\item{dept}{Department code}
}
\description{
Write course data to cache
}
