% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{write_json_cache}
\alias{write_json_cache}
\title{Write JSON data to cache}
\usage{
write_json_cache(data, cache_dir, filename)
}
\arguments{
\item{data}{Data frame or list to cache}

\item{cache_dir}{Base cache directory}

\item{filename}{Filename without extension}
}
\description{
Write JSON data to cache
}
\keyword{internal}
