% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear cache for specific data or entirely}
\usage{
clear_cache(name = NULL, year = NULL, cache_dir = NULL)
}
\arguments{
\item{name}{Department code (optional)}

\item{year}{Academic year (optional)}

\item{cache_dir}{Base cache directory (optional)}
}
\value{
Invisibly returns TRUE if successful, FALSE otherwise
}
\description{
Clear cache for specific data or entirely
}
