% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{extract_schedule_data}
\alias{extract_schedule_data}
\title{Extract schedule data from a section including instructors and their roles}
\usage{
extract_schedule_data(section)
}
\arguments{
\item{section}{XML node for a section}
}
\value{
Tibble with schedule and instructor information
}
\description{
Extract schedule data from a section including instructors and their roles
}
\keyword{internal}
