% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_all_courses}
\alias{fetch_all_courses}
\title{Fetch and process courses for multiple departments}
\usage{
fetch_all_courses(departments = NULL, cache_dir = NULL)
}
\arguments{
\item{departments}{Character vector of department codes}

\item{cache_dir}{Directory to cache results}
}
\value{
A list of data frames containing course information
}
\description{
Fetch and process courses for multiple departments
}
