% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_department_courses}
\alias{fetch_department_courses}
\title{Fetch courses for a specific department with progress reporting}
\usage{
fetch_department_courses(name, cache_dir = NULL, p = NULL)

fetch_department_courses(name, cache_dir = NULL, p = NULL)
}
\arguments{
\item{name}{Department code}

\item{cache_dir}{Base cache directory}

\item{p}{Progress handler}
}
\value{
Data frame of parsed course information

Data frame of parsed course information
}
\description{
Fetch courses for a specific department with progress reporting

Fetch courses for a specific department with progress reporting
}
