% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_department_courses}
\alias{fetch_department_courses}
\title{Fetch courses for a specific department}
\usage{
fetch_department_courses(name, cache_dir = NULL)
}
\arguments{
\item{name}{Department code}

\item{cache_dir}{Directory to cache results}
}
\value{
XML content of courses
}
\description{
Fetch courses for a specific department
}
