% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{fetch_department_courses_raw}
\alias{fetch_department_courses_raw}
\title{Fetch raw course data for a department from API}
\usage{
fetch_department_courses_raw(name, year = NULL)
}
\arguments{
\item{name}{Department code}

\item{year}{Academic year in format YYYYYYYY or NULL for current year}
}
\value{
Raw XML content
}
\description{
Fetch raw course data for a department from API
}
\keyword{internal}
