% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_departments}
\alias{fetch_departments}
\title{Fetch department list from Stanford ExploreCourses}
\usage{
fetch_departments(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{Base cache directory}
}
\value{
A data frame containing department information
}
\description{
Fetch department list from Stanford ExploreCourses
}
