% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_academic_year}
\alias{generate_academic_year}
\title{Generate academic year string}
\usage{
generate_academic_year(start_year)
}
\arguments{
\item{start_year}{Starting year as YYYY}
}
\value{
Academic year string in YYYYYYYY format
}
\description{
Generate academic year string
}
\examples{
generate_academic_year(2023)  # Returns "20232024"
}
