% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{list_cache}
\alias{list_cache}
\title{List cached files}
\usage{
list_cache(year = NULL, cache_dir = NULL)
}
\arguments{
\item{year}{Academic year (optional)}

\item{cache_dir}{Base cache directory (optional)}
}
\value{
Character vector of cached file names (without extensions)
}
\description{
List cached files
}
