% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{print.explorecourses_cache_list}
\alias{print.explorecourses_cache_list}
\title{Print method for cache list}
\usage{
\method{print}{explorecourses_cache_list}(x, ...)
}
\arguments{
\item{x}{Object of class explorecourses_cache_list}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for cache list
}
