% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_courses_xml}
\alias{process_courses_xml}
\title{Process courses XML data into a data frame}
\usage{
process_courses_xml(xml_doc, department)
}
\arguments{
\item{xml_doc}{XML document (xml2 doc object)}

\item{department}{Department code}
}
\value{
Data frame of courses with detailed schedule and instructor information
}
\description{
Process courses XML data into a data frame
}
\keyword{internal}
