% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{write_xml_cache}
\alias{write_xml_cache}
\title{Write XML data to cache}
\usage{
write_xml_cache(content, cache_dir, dept)
}
\arguments{
\item{content}{XML content}

\item{cache_dir}{Base cache directory}

\item{dept}{Department code}
}
\description{
Write XML data to cache
}
\keyword{internal}
